Move Convert methods from XML_Engine to XML_Adapter #440
