+++ XML_Adapter/Convert/GBXML/Architecture/Level.cs
@@ -25,51 +25,56 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+
 using BHE = BH.oM.Environment.Elements;
+using BHA = BH.oM.Architecture.Elements;
 using BHX = BH.oM.XML;
 using BHG = BH.oM.Geometry;
 
 using BH.Engine.Geometry;
-
+using BH.Engine.Environment;
 using System.ComponentModel;
 using BH.oM.Reflection.Attributes;
 
-namespace BH.Engine.XML
+namespace BH.Adapter.XML
 {
     public static partial class Convert
     {
-        [Description("Get the GBXML representation of a BHoM Point")]
-        [Input("pt", "The BHoM Geometry Point to convert into a GBXML Cartesian Point")]
-        [Output("cartesianPoint", "The GBXML representation of a BHoM Point")]
-        public static BHX.CartesianPoint ToGBXML(this BHG.Point pt)
+        [Description("Returns a gbXML BuildingStorey represention of a BHoM level and spaces")]
+        [Input("level", "A BHoM level to find the storey for")]
+        [Input("spaces", "A list of BHoM spaces that sits on the given level")]
+        [Output("buildingStorey", "The gbXML building storey")]
+        public static BHX.BuildingStorey ToGBXML(this BHG.SettingOut.Level level, List<List<BHE.Panel>> spaces = null)
         {
-            BHX.CartesianPoint cartPoint = new BHX.CartesianPoint();
-            List<string> coord = new List<string>();
+            BHX.BuildingStorey storey = new BHX.BuildingStorey();
 
-            coord.Add(Math.Round(pt.X, 4).ToString());
-            coord.Add(Math.Round(pt.Y, 4).ToString());
-            coord.Add(Math.Round(pt.Z, 4).ToString());
+            if (spaces != null)
+            {
+                BHG.Polyline storeyGeometry = level.StoreyGeometry(spaces);
+                if (storeyGeometry != null)
+                    storey.PlanarGeometry.PolyLoop = storeyGeometry.ToGBXML();
+            }
 
-            cartPoint.Coordinate = coord.ToArray();
+            storey.PlanarGeometry.ID = "level-planar-geometry-" + Guid.NewGuid().ToString().Replace("-", "").Substring(0, 10);
+            storey.Name = level.Name;
+            storey.ID = "Level-" + level.Name.Replace(" ", "").ToLower();
+            storey.Level = (float)level.Elevation;
 
-            return cartPoint;
+            return storey;
         }
 
-        [Description("Get the BHoM representation of a GBXML Point")]
-        [Input("pt", "The GBXML Cartesian Point to convert into a BHoM Geometry Point")]
-        [Output("bhomPoint", "The BHoM representation of a GBXML Point")]
-        public static BHG.Point FromGBXML(this BHX.CartesianPoint pt)
+        [Description("Returns a BHoM Level represention of a gbXML storey")]
+        [Input("storey", "Set a gbXML storey to get the Level from")]
+        [Output("level", "The BHoM level")]
+        public static BHG.SettingOut.Level FromGBXML(this BHX.BuildingStorey storey)
         {
-            BHG.Point bhomPt = new BHG.Point();
-            try
-            {
-                bhomPt.X = (pt.Coordinate.Length >= 1 ? System.Convert.ToDouble(pt.Coordinate[0]) : 0);
-                bhomPt.Y = (pt.Coordinate.Length >= 2 ? System.Convert.ToDouble(pt.Coordinate[1]) : 0);
-                bhomPt.Z = (pt.Coordinate.Length >= 3 ? System.Convert.ToDouble(pt.Coordinate[2]) : 0);
-            }
-            catch { }
+            BHG.SettingOut.Level level = new BHG.SettingOut.Level();
 
-            return bhomPt;
+            level.Name = storey.Name;
+            level.Elevation = storey.Level;
+
+            return level;
         }
     }
 }
+
