+++ XML_Adapter/Convert/GBXML/Environment/Construction.cs
@@ -41,7 +41,7 @@ using BHM = BH.oM.Physical.Materials;
 using System.ComponentModel;
 using BH.oM.Reflection.Attributes;
 
-namespace BH.Engine.XML
+namespace BH.Adapter.XML
 {
     public static partial class Convert
     {
