+++ XML_Adapter/Convert/GBXML/Environment/ElementType.cs
@@ -39,7 +39,7 @@ using BH.oM.XML.Enums;
 using System.ComponentModel;
 using BH.oM.Reflection.Attributes;
 
-namespace BH.Engine.XML
+namespace BH.Adapter.XML
 {
     public static partial class Convert
     {
