+++ XML_Adapter/Convert/GBXML/Environment/Roughness.cs
@@ -33,7 +33,7 @@ using BHX = BH.oM.XML;
 using System.ComponentModel;
 using BH.oM.Reflection.Attributes;
 
-namespace BH.Engine.XML
+namespace BH.Adapter.XML
 {
     public static partial class Convert
     {
