+++ XML_Adapter/Convert/GBXML/Geometry/Polyline.cs
@@ -35,7 +35,7 @@ using BH.Engine.Geometry;
 using System.ComponentModel;
 using BH.oM.Reflection.Attributes;
 
-namespace BH.Engine.XML
+namespace BH.Adapter.XML
 {
     public static partial class Convert
     {
